# packages/web/src/lib/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SAAKSHI Registry API",
    "description": "Camera registry, onboarding and export. Model 1 names three onboarding paths — bulk import, manual entry and API — and all three are served here.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/cameras/{id}/trust": {
      "get": {
        "summary": "Current trust score, the breakdown that explains it, and the recent trend",
        "tags": [
          "trust"
        ],
        "parameters": [
          {
            "schema": {
              "default": 7,
              "type": "integer",
              "minimum": 1,
              "maximum": 90
            },
            "in": "query",
            "name": "days",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cameraId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "externalId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "score": {
                      "nullable": true,
                      "type": "number"
                    },
                    "band": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "trusted",
                        "degraded",
                        "untrusted",
                        "dead"
                      ]
                    },
                    "checkedAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "breakdown": {
                      "type": "object",
                      "properties": {
                        "signals": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "signal": {
                                "type": "string"
                              },
                              "raw": {
                                "anyOf": [
                                  {
                                    "type": "number"
                                  },
                                  {
                                    "type": "boolean"
                                  }
                                ],
                                "nullable": true
                              },
                              "quality": {
                                "nullable": true,
                                "type": "number"
                              },
                              "weight": {
                                "type": "number"
                              },
                              "points": {
                                "type": "number"
                              },
                              "maxPoints": {
                                "type": "number"
                              },
                              "applicable": {
                                "type": "boolean"
                              },
                              "note": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "signal",
                              "raw",
                              "quality",
                              "weight",
                              "points",
                              "maxPoints",
                              "applicable",
                              "note"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "excluded": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "signal": {
                                "type": "string"
                              },
                              "reason": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "signal",
                              "reason"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "weightsVersion": {
                          "nullable": true,
                          "type": "number"
                        },
                        "pointsTotal": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "signals",
                        "excluded",
                        "weightsVersion",
                        "pointsTotal"
                      ],
                      "additionalProperties": false
                    },
                    "trend": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "bucket": {
                            "type": "string"
                          },
                          "score": {
                            "nullable": true,
                            "type": "number"
                          },
                          "band": {
                            "nullable": true,
                            "type": "string",
                            "enum": [
                              "trusted",
                              "degraded",
                              "untrusted",
                              "dead"
                            ]
                          },
                          "checks": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "reachableChecks": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "bucket",
                          "score",
                          "band",
                          "checks",
                          "reachableChecks"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "cameraId",
                    "externalId",
                    "name",
                    "score",
                    "band",
                    "checkedAt",
                    "breakdown",
                    "trend"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "allowed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "allowed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "allowed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}